azure-data-pipeline/functions/DataCollector/requirements.txt
azure-functions==1.17.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-iot-hub==2.3.0
azure-identity==1.15.0
python-dotenv==1.0.0
